Move FFmpeg pointer wrappers into utils header
