+++ src/core/engine/ffmpeg/ffmpegresampler.h
@@ -21,10 +21,24 @@
 
 #include <core/engine/audiobuffer.h>
 
+#if defined(__GNUG__)
+#pragma GCC diagnostic push
+#pragma GCC diagnostic ignored "-Wold-style-cast"
+#elif defined(__clang__)
+#pragma clang diagnostic push
+#pragma clang diagnostic ignored "-Wold-style-cast"
+#endif
+
 extern "C"
 {
 #include <libswresample/swresample.h>
 }
+
+#if defined(__GNUG__)
+#pragma GCC diagnostic pop
+#elif defined(__clang__)
+#pragma clang diagnostic pop
+#endif
 
 namespace Fooyin {
 struct SwrContextDeleter
